Update main.py
Fixed bug that caused multiple buttons that overlay to operate at the same time, ex Back and Quit, clicking back would also count as clicking Quit and exit the game entirely. Created a separate value called Click, and checked each time if it was false to allow the button's process to continue.